+++ src/main/java/de/bnder/taskmanager/lists/DiscordbotlistCOM.java
@@ -11,7 +11,7 @@ public class DiscordbotlistCOM {
     private static final String baseURL = "https://discordbotlist.com/api/v1";
     private static final String apiKey = Main.dotenv.get("DISCORDBOTLIST.COM_API_KEY") != null ? Main.dotenv.get("DISCORDBOTLIST.COM_API_KEY") : System.getenv("DISCORDBOTLIST.COM_API_KEY");
 
-    public static void sendServerCount(int serverCount, String botID) throws IOException {
+    public static void sendServerCount(long serverCount, String botID) throws IOException {
         System.out.println("Updating Servers on " + baseURL);
         final Connection.Response response = Jsoup.connect(baseURL + "/bots/"+ botID + "/stats")
                 .header("Authorization", apiKey)
