+++ src/main/java/de/bnder/taskmanager/lists/InfinitybotlistCOM.java
@@ -26,7 +26,7 @@ public class InfinitybotlistCOM {
     private static final String baseURL = "https://infinitybotlist.com/api";
     private static final String apiKey = Main.dotenv.get("INFINITYBOTLIST.COM_API_KEY") != null ? Main.dotenv.get("INFINITYBOTLIST.COM_API_KEY") : System.getenv("INFINITYBOTLIST.COM_API_KEY");
 
-    public static void sendServerCount(int serverCount, String botID) throws IOException {
+    public static void sendServerCount(long serverCount, String botID) throws IOException {
         System.out.println("Updating Servers on " + baseURL);
         final Connection.Response response = Jsoup.connect(baseURL + "/bots/" + botID)
                 .header("authorization", apiKey)
