+++ src/main/java/de/bnder/taskmanager/lists/TopGG.java
@@ -11,7 +11,7 @@ public class TopGG {
     private static final String baseURL = "https://top.gg/api";
     private static final String apiKey = Main.dotenv.get("TOP.GG_API_KEY") != null ? Main.dotenv.get("TOP.GG_API_KEY") : System.getenv("TOP.GG_API_KEY");
 
-    public static void sendServerCount(int serverCount, String botID) throws IOException {
+    public static void sendServerCount(long serverCount, String botID) throws IOException {
         System.out.println("Updating Servers on " + baseURL);
         final Connection.Response response = Jsoup.connect(baseURL + "/bots/" + botID + "/stats")
                 .header("Authorization", apiKey)
